backend/src/middleware/jwt_auth.py
from fastapi import Request, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from jose import JWTError, jwt
from datetime import datetime, timedelta
from typing import Optional
import uuid
from ..config import settings


class JWTBearer(HTTPBearer):
    def __init__(self, auto_error: bool = True):
        super(JWTBearer, self).__init__(auto_error=auto_error)

    async def __call__(self, request: Request):
        credentials: HTTPAuthorizationCredentials = await super(JWTBearer, self).__call__(request)

        if credentials:
            if not credentials.scheme == "Bearer":
                raise HTTPException(
                    status_code=status.HTTP_401_UNAUTHORIZED,
                    detail="Invalid authentication scheme."
                )

            token = credentials.credentials
            user_id = self.verify_jwt(token)

            if not user_id:
                raise HTTPException(
                    status_code=status.HTTP_401_UNAUTHORIZED,
                    detail="Invalid or expired token."
                )

            # Add user_id to request state for use in endpoints
            request.state.user_id = user_id
            return credentials.credentials
        else:
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Invalid authorization code."
            )

    def verify_jwt(self, token: str) -> Optional[uuid.UUID]:
        try:
            payload = jwt.decode(
                token,
                settings.jwt_secret,
                algorithms=["HS256"]
            )

            user_id: str = payload.get("user_id")
            exp: int = payload.get("exp")

            if user_id is None or exp is None:
                return None

            # Check if token is expired
            if datetime.fromtimestamp(exp) < datetime.utcnow():
                return None

            return uuid.UUID(user_id)
        except JWTError:
            return None


def create_access_token(user_id: uuid.UUID) -> str:
    """
    Create a new JWT access token
    """
    expire = datetime.utcnow() + timedelta(hours=24)  # Default to 24 hours

    to_encode = {
        "user_id": str(user_id),
        "exp": expire.timestamp(),
        "iat": datetime.utcnow().timestamp(),
        "type": "access"
    }

    encoded_jwt = jwt.encode(to_encode, settings.jwt_secret, algorithm="HS256")
    return encoded_jwt