backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.7.0